Upgrade to inertia.js 1.0.0

+++ app/Actions/Auth/GeneratePassword.php
@@ -6,16 +6,16 @@ class GeneratePassword
 {
     public function __invoke()
     {
-        $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', ceil(12 / strlen($x)))), 1, 20);
+        $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', (int) ceil(12 / strlen($x)))), 1, 20);
 
         // Ensure it has numbers
         while (! preg_match('/[0-9]/', $str)) {
-            $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', ceil(12 / strlen($x)))), 1, 20);
+            $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', (int) ceil(12 / strlen($x)))), 1, 20);
         }
 
         // Ensure it has letters
         while (! preg_match('/[a-zA-Z]/', $str)) {
-            $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', ceil(12 / strlen($x)))), 1, 20);
+            $str = substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+-=', (int) ceil(12 / strlen($x)))), 1, 20);
         }
 
         return $str;
